pyproject.toml
[project]
name = "llm-city-assistant"
version = "0.1.0"
description = "Интеллектуальный агент «Городской помощник» для Санкт-Петербурга"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.5",
    "gigachat>=0.1.43",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-gigachat>=0.3.13a1",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langsmith>=0.4.49",
    "numpy>=2.3.5",
    "pydantic>=2.12.5",
    "pymorphy2>=0.9.1",
    "pymorphy2-dicts-ru>=2.4.417127.4579844",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "setuptools>=80.9.0",
    "streamlit>=1.51.0",
    "streamlit-authenticator>=0.4.2",
    "structlog>=25.5.0",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "apscheduler>=3.11.1",
    "huggingface-hub>=0.36.0",
    "ipykernel>=7.1.0",
    "langgraph-cli[inmem]>=0.4.7",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.6",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
types = [
    "types-pyyaml>=6.0.12.20250915",
]
